// src/types/partials/Address.ts
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TPeriod } from '../partials/Period';

export type TAddress = {
    id?: String;
    extension?: TExtension[];
    use?: String;
    type?: String;
    text?: String;
    line?: String[];
    city?: String;
    district?: String;
    state?: String;
    postalCode?: String;
    country?: String;
    period?: TPeriod;
};